telegram/requirements.txt
python-telegram-bot==12.1.1
PyYAML==5.1.1
redis==3.1.0
rq==0.13.0